fix(jangar): rename torghut visuals to charts and restore ta auth
