--- argocd/applications/jangar/clickhouse-auth.yaml
@@ -1,17 +1,0 @@
----
-apiVersion: bitnami.com/v1alpha1
-kind: SealedSecret
-metadata:
-  name: jangar-clickhouse-auth
-  namespace: jangar
-  annotations:
-    argocd.argoproj.io/ignore-healthcheck: "true"
-spec:
-  encryptedData:
-    password: AgBbGFf30eZlWTvnPPamhtoSi0LY/ey2t4hq8BFDtKeNIgnEEuiCeDDtcWk8V8Ci6XxNLiOQtD6zeh1gC5NCa6f9JVM/0KT5mL8H7UCfM3OO/Aau2nIvLIOqmkfIgrCo2qcNoOT5OJQCL2ED/r7iPkwGkPzciEhS1qByFY5tNp0VbRSjxeMbFwHdgMxwovV7yaMhYVTXG78V8emqY1lxsh+qc8lg99gNqn7x/9v8uHhYKktOFODlaP6uZmrbM7FlB9RwKeew2vvMsNgsbsvLjPjV0XrDfYB6ON5xzNlxaQ7bt7Do8xb3LlLrUpdLuAtmaED0wyN1dYRub2vE7Unqo+pcHbv2zGnw6V2CfXSVW16CQTytDiYD2+rLZrK7kOXAF3wNzCAxm4ruQXDyRuBM+yPpAweYZihxzMMeDJ5o/DA/gazWhD9mMO12DP6hxgczJHECfHbQPcgBbJaQ7icZMtQR5oZR8eMS8fsdn5Jsyv7qd+l45FPDNUv1wo8fnKvjgIrdj+2fxL7YuImEz3Ud4E3qE8VObw6qOCo/hVk64FG2ItaIzu4P8iVTyyx70+UVukNkQ/rvleffOlz1HjDEa+XaI2FKEQYOwSe1jGckx6j7SH5cBFXynN1xKK9d2a5Lo/i0/ZzT9SIwlxBvgoqyss+DCs004ysDjvKc51yvnJ0t73aRboN2fVkXFntdggk5vEFFd63+h0iOoNqhYzRRy5CGnRqVPnVAWWZdx5V19SS0uA==
-    username: AgBRkESspX9jZKWmIjnZDsAW32jkdsOz2WyNphPlYJjM+YmJmNZ+/ovFGqtHhh4WJ1PDvuqGFQRUvaLiSkzlfa5xLKwSlYCfgTA6EvEUjcqheKlwHsrXv4pOBhXdjnNzPzMmyMv9gnUBvQ3W9l741G3/kH8OBcvL+9O3THaouuCYO9tI3Kqr0i6WwA0Fg8jokNnFL/CVQbhN9F2IESyD3/JXWfUNk1oL2j11WNhihdKLTTH8ZPLs39b2zNbEaDdxIPy6FC7tBSuvy61YK31GtUh9NMDR6sl901nTz9ZiH7VFg4RI9PpCS3ZiI/IsbPnJYf3T/D9IJYpVfX4HIUyb62tBq96LejDf1wGCZUNWTENj2/vitpClVSpqIostDScv+UJqNisQur6+gnrUr/PDEuct/SQuv7xnviJUtPj63Eg6BXsKUlQnQ7a7Gn9Bf/qZ5G+SPxLEG22M8x9u3ajfyOA2lRjIMQDHcs7xhFddQXS3rrbLqh+2XM694sfEffJn5n7IssmhQ7YkNEtzTGCZ2ttPmjLnTW0GtMo6VktQ+o95HS193o9MImjheHLHyK+yvSKbnU6ocL8/Wop4gNK65EdTe1u1KuAYBCY2VPTuXOMJVjYRFibdl5gCqKf9J7EzsNJvnfroHETG6dU+tNDdD41zKUDO+R866DaeHCxJv2GH+b4C24kFxnTdaKGfb0sPGd0sNQfRHyqc
-  template:
-    metadata:
-      name: jangar-clickhouse-auth
-      namespace: jangar
-    type: Opaque
