+++ argocd/applications/torghut/clickhouse/sealed-secret.yaml
@@ -24,6 +24,6 @@ spec:
         app.kubernetes.io/part-of: torghut
       annotations:
         reflector.v1.k8s.emberstack.com/reflection-allowed: "true"
-        reflector.v1.k8s.emberstack.com/reflection-allowed-namespaces: "pgadmin"
+        reflector.v1.k8s.emberstack.com/reflection-allowed-namespaces: "pgadmin,jangar"
         reflector.v1.k8s.emberstack.com/reflection-auto-enabled: "true"
-        reflector.v1.k8s.emberstack.com/reflection-auto-namespaces: "pgadmin"
+        reflector.v1.k8s.emberstack.com/reflection-auto-namespaces: "pgadmin,jangar"
