+++ services/jangar/src/components/app-sidebar.tsx
@@ -97,7 +97,7 @@ const apiNav = [
 
 const torghutNav = [
   { to: '/torghut/symbols', label: 'Symbols', icon: IconList },
-  { to: '/torghut/visuals', label: 'Visuals', icon: IconChartCandle },
+  { to: '/torghut/charts', label: 'Charts', icon: IconChartCandle },
   { to: '/torghut/trading', label: 'Trading', icon: IconChartLine },
 ] as const
